refactor: move fetch package to cmd/motr
