+++ cmd/motr/main.go
@@ -5,8 +5,8 @@ package main
 
 import (
 	"github.com/labstack/echo/v4"
+	"github.com/onsonr/sonr/cmd/motr/fetch"
 	"github.com/onsonr/sonr/internal/ctx"
-	"github.com/onsonr/sonr/internal/dwn/fetch"
 	"github.com/onsonr/sonr/workers/routes"
 )
